fix: 리뷰어 님 피드백 반영
- CalculatorExpression이 값 전달 객체 이므로 dto 패키지로 이동
- StringAddCalculator 내 validate 로직을 별도 예외를 throw 하도록 수정
- CaculateNumber 내 유효성 검사 로직을 별도 private 메소드로 분리
- CalculateNumbersFactory가 하던 팩토리 로직을 CalculateNumbers에게 위임 및 관련 테스트 코드 수정
