+++ src/main/java/calculator/domain/StringAddCalculator.java
@@ -1,23 +1,28 @@
 package calculator.domain;
 
+import calculator.dto.CalculatorExpression;
 import calculator.util.CalculatorExpressionFactory;
-import calculator.util.CalculatorNumbersFactory;
 
 import java.util.regex.Pattern;
 
 public class StringAddCalculator {
     private static final int DEFAULT_RESULT = 0;
+    private static final String NPE_MESSAGE = "값이 null 이거나 비어 있습니다.";
 
     private static final Pattern pattern = Pattern.compile("//(.)\n(.*)");
+
 
     private StringAddCalculator() {
 
     }
 
     public static int calculate(String expression) {
-        if (validate(expression)) {
+        try {
+            validate(expression);
+        } catch (NullPointerException e) {
             return DEFAULT_RESULT;
         }
+
         CalculatorExpression calculatorExpression = CalculatorExpressionFactory.of(expression, pattern);
 
         String separator = calculatorExpression.getSeparator();
@@ -27,13 +32,15 @@ public class StringAddCalculator {
         return calculatorNumbers.sum();
     }
 
-    private static boolean validate(String expression) {
-        return expression == null || expression.isEmpty();
+    private static void validate(String expression) {
+        if(expression == null || expression.isEmpty()) {
+            throw new NullPointerException(NPE_MESSAGE);
+        }
     }
 
     private static CalculatorNumbers extractCalculateNumbers(String separator, String formula) {
         String[] splitResults = formula.split(separator);
 
-        return CalculatorNumbersFactory.from(splitResults);
+        return CalculatorNumbers.from(splitResults);
     }
 }
