+++ src/main/java/calculator/dto/CalculatorExpression.java
@@ -1,4 +1,4 @@
-package calculator.domain;
+package calculator.dto;
 
 public class CalculatorExpression {
