+++ src/test/java/calculator/util/CalculatorExpressionFactoryTest.java
@@ -1,6 +1,6 @@
 package calculator.util;
 
-import calculator.domain.CalculatorExpression;
+import calculator.dto.CalculatorExpression;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
